Create a forest matrice for legal entity from insee code

Generates a forest matrice used to store general forest information (e.g., IDENTIFIANT, PROPRIETAIRE) and cadastral attributes (CODE_INSEE, PREFIXE, SECTION, NUMERO, LIEU_DIT) for legal entity.

Usage

get_legal_entity(x, cache = NULL, verbose = TRUE)

Arguments

x

character; Code(s) INSEE or code(s) department (see happign::com_2025 or happign::dep_2025)

cache

character; Storage directory. Defaults to the user cache directory (see tools::R_user_dir()).

verbose

logical If TRUE, display messages.

Details

Matrice from get_legal_entity add tx_boisee information which is the percentage of land considered as forest :

"L": Landes

"LB": Landes Boisees

"B": Bois

"BF": Futaies Feuillues

"BM": Futaies Mixtes

"BO": Oseraies

"BP": Peupleraies

"BR": Futaies Resineuses

"BS": Taillis sous Futaies

"BT": Taillis Simples

Examples

if (FALSE) { # \dontrun{
pm <- get_legal_entity(c("29158", "08185"))
} # }